Revert "Change to double"
This reverts commit 4690ffb7e39352c6b276b9f09e12a13fb1be7304.

+++ tree.h
@@ -1,5 +1,5 @@
-double tree[3][2];
-double trunk[4][2];
+int tree[3][2];
+int trunk[4][2];
 
 void initTree();
 void drawTree();
